Warnings Draft Implementation
The product of adding warnings to VisualGit in the case of users wanting to Exit / Sign Out / Pull if there are commits that have not been commited and if the repository has not been pushed.

Warnings Draft Implementation

The product of adding warnings to VisualGit in the case of users wanting to Exit / Sign Out / Pull if there are commits that have not been commited and if the repository has not been pushed.

Added option to reload

As opposed to only exiting the application when prompted; added the feature of allowing users to choose between reloading or exiting application, along with a back button.

Commit after Revert; grammar / commenting changed

Ready for another pull request after fixing up all previous errors and reverting failed pull request.